split stuff single file to module; fixes needed
